+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - feat: support for workspace trust
 - feat: support config for HTML formatting print width by `volar.formatting.printWidth` option ([#321](https://github.com/johnsoncodehk/volar/issues/321))
 - feat: support for typescript `updateImportsOnFileMove` setting to disable prompt ([#332](https://github.com/johnsoncodehk/volar/issues/332))
+- feat: add "Show in Browser" button to component preview
 - fix: component preview not working on windows
 - fix: delete file can't trigger related scripts diagnostics update
 - fix: disable component tag type-checking to avoid some unable fix edge cases ([#333](https://github.com/johnsoncodehk/volar/issues/333))